OPL-71 Moved business logic to services
